+++ lib/openstudio-workflow.rb
@@ -17,11 +17,12 @@
 #  Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
 ######################################################################
 
-require 'aasm'
 require 'pp'
+require 'rubyXL'
 require 'multi_json'
 require 'colored'
 require 'fileutils'
+require 'securerandom'
 require 'json' # needed for a single pretty generate call
 require 'pathname'
 
@@ -35,6 +36,7 @@ require 'openstudio/workflow/version'
 require 'openstudio/workflow/multi_delegator'
 require 'openstudio/workflow/run'
 require 'openstudio/workflow/jobs/lib/apply_measures'
+require 'openstudio/workflow/time_logger'
 
 begin
   require 'openstudio'
@@ -48,10 +50,10 @@ end
 class String
   def snake_case
     gsub(/::/, '/')
-        .gsub(/([A-Z]+)([A-Z][a-z])/, '\1_\2')
-        .gsub(/([a-z\d])([A-Z])/, '\1_\2')
-        .tr(' -', '__')
-        .downcase
+      .gsub(/([A-Z]+)([A-Z][a-z])/, '\1_\2')
+      .gsub(/([a-z\d])([A-Z])/, '\1_\2')
+      .tr(' -', '__')
+      .downcase
   end
 end
 
@@ -66,11 +68,11 @@ module OpenStudio
 
       # Convert various paths to absolute paths
       if options[:adapter_options] && options[:adapter_options][:mongoid_path] &&
-          (Pathname.new options[:adapter_options][:mongoid_path]).absolute? == false
+         (Pathname.new options[:adapter_options][:mongoid_path]).absolute? == false
         options[:adapter_options][:mongoid_path] = File.expand_path options[:adapter_options][:mongoid_path]
       end
       if options[:analysis_root_path] &&
-          (Pathname.new options[:analysis_root_path]).absolute? == false
+         (Pathname.new options[:analysis_root_path]).absolute? == false
         options[:analysis_root_path] = File.expand_path options[:analysis_root_path]
       end
       unless (Pathname.new run_directory).absolute?
@@ -80,6 +82,38 @@ module OpenStudio
       adapter = load_adapter adapter_name, options[:adapter_options]
       run_klass = OpenStudio::Workflow::Run.new(adapter, run_directory, options)
       # return the run class
+      run_klass
+    end
+
+    # predefined method that simply runs EnergyPlus in the specified directory. It does not apply any workflow steps
+    # such as preprocessing / postprocessing.
+    # The directory must have the IDF and EPW file in the folder. The simulations will run in the directory/run path
+    def run_energyplus(adapter_name, run_directory, options = {})
+      unless (Pathname.new run_directory).absolute?
+        # relative to wherever you are running the script
+        run_directory = File.expand_path run_directory
+      end
+
+      transitions = [
+        { from: :queued, to: :preflight },
+        { from: :preflight, to: :energyplus },
+        { from: :energyplus, to: :finished }
+      ]
+
+      states = [
+        { state: :queued, options: { initial: true } },
+        { state: :preflight, options: { after_enter: :run_preflight } },
+        { state: :energyplus, options: { after_enter: :run_energyplus } },
+        { state: :finished },
+        { state: :errored }
+      ]
+      options = {
+        transitions: transitions,
+        states: states
+      }
+
+      adapter = load_adapter adapter_name, options[:adapter_options]
+      run_klass = OpenStudio::Workflow::Run.new(adapter, run_directory, options)
       run_klass
     end