+++ lib/openstudio/workflow/jobs/run_energyplus/run_energyplus.rb
@@ -21,7 +21,7 @@ class RunEnergyplus
   # Initialize
   # param directory: base directory where the simulation files are prepared
   # param logger: logger object in which to write log messages
-  def initialize(directory, logger, adapter, options = {})
+  def initialize(directory, logger, time_logger, adapter, options = {})
     energyplus_path = nil
     if /cygwin|mswin|mingw|bccwin|wince|emx/ =~ RUBY_PLATFORM
       energyplus_path = 'C:/EnergyPlus-8-1-0'
@@ -39,6 +39,7 @@ class RunEnergyplus
     @run_directory = "#{@directory}/run"
     @adapter = adapter
     @logger = logger
+    @time_logger = time_logger
     @results = {}
 
     @logger.info "#{self.class} passed the following options #{@options}"
@@ -51,22 +52,54 @@ class RunEnergyplus
     # Ensure that the directory is created (but it should already be at this point)
     FileUtils.mkdir_p(@run_directory)
 
+    # if the weather file is already in the directory, then just use that weather file
+    weather_file_name = nil
+    weather_files = Dir["#{@directory}/*.epw"]
+    if weather_files.size > 1
+      @logger.info 'Multiple weather files in the directory. Will rely on the weather file name in the openstudio model'
+    elsif weather_files.size == 1
+      weather_file_name = weather_files.first
+    end
+
     # verify that the OSM, IDF, and the Weather files are in the run directory as the 'in.*' format
-    if @options[:run_openstudio][:weather_filename] && File.exist?(@options[:run_openstudio][:weather_filename])
+    if !weather_file_name &&
+       @options[:run_openstudio] &&
+       @options[:run_openstudio][:weather_filename] &&
+       File.exist?(@options[:run_openstudio][:weather_filename])
+      weather_file_name = @options[:run_openstudio][:weather_filename]
+    end
+
+    if weather_file_name
       # verify that it is named in.epw
-      unless File.basename(@options[:run_openstudio][:weather_filename]).downcase == 'in.idf'
-        FileUtils.copy(@options[:run_openstudio][:weather_filename], "#{@run_directory}/in.epw")
-      end
+      @logger.info "Weather file for EnergyPlus simulation is #{weather_file_name}"
+      FileUtils.copy(weather_file_name, "#{@run_directory}/in.epw")
     else
       fail "EPW file not found or not sent to #{self.class}"
     end
 
+    # check if the run folder has an IDF. If not then check if the parent folder does.
+    idf_file_name = nil
+    if File.exist?("#{@run_directory}/in.idf")
+      @logger.info 'IDF (in.idf) already exists in the run directory'
+    else
+      # glob for idf at the directory level
+      idfs = Dir["#{@directory}/*.idf"]
+      if idfs.size > 1
+        @logger.info 'Multiple IDF files in the directory. Cannot continue'
+      elsif idfs.size == 1
+        idf_file_name = idfs.first
+      end
+    end
+
     # Need to check the in.idf and in.osm
     # FileUtils.copy(options[:osm], "#{@run_directory}/in.osm")
-    # FileUtils.copy(options[:idf], "#{@run_directory}/in.idf")
+    if idf_file_name
+      FileUtils.copy(idf_file_name, "#{@run_directory}/in.idf")
+    end
 
     # can't create symlinks because the /vagrant mount is actually a windows mount
     @logger.info "Copying EnergyPlus files to run directory: #{@run_directory}"
+    @time_logger.start("Copying EnergyPlus files")
     FileUtils.copy("#{@options[:energyplus_path]}/libbcvtb.so", "#{@run_directory}/libbcvtb.so")
     FileUtils.copy("#{@options[:energyplus_path]}/libepexpat.so", "#{@run_directory}/libepexpat.so")
     FileUtils.copy("#{@options[:energyplus_path]}/libepfmiimport.so", "#{@run_directory}/libepfmiimport.so")
@@ -75,8 +108,11 @@ class RunEnergyplus
     FileUtils.copy("#{@options[:energyplus_path]}/ExpandObjects", "#{@run_directory}/ExpandObjects")
     FileUtils.copy("#{@options[:energyplus_path]}/EnergyPlus", "#{@run_directory}/EnergyPlus")
     FileUtils.copy("#{@options[:energyplus_path]}/Energy+.idd", "#{@run_directory}/Energy+.idd")
+    @time_logger.stop("Copying EnergyPlus files")
 
+    @time_logger.start("Running EnergyPlus")
     @results = call_energyplus
+    @time_logger.stop("Running EnergyPlus")
 
     @results
   end
@@ -105,15 +141,41 @@ class RunEnergyplus
 
       # create stdout
       File.open('stdout-energyplus', 'w') do |file|
-        IO.popen('./EnergyPlus') do |io|
+        IO.popen('./EnergyPlus + 2>&1') do |io|
           while (line = io.gets)
             file << line
           end
         end
       end
+      r = $?
+
+      @logger.info "System call to EnergyPlus returned #{r}"
+
+      paths_to_rm = []
+      paths_to_rm << Pathname.glob("#{@run_directory}/*.ini")
+      paths_to_rm << Pathname.glob("#{@run_directory}/*.so")
+      paths_to_rm << Pathname.glob("#{@run_directory}/*.idd")
+      paths_to_rm << Pathname.glob("#{@run_directory}/ExpandObjects")
+      paths_to_rm << Pathname.glob("#{@run_directory}/EnergyPlus")
+      paths_to_rm << Pathname.glob("#{@run_directory}/packaged_measures")
+      paths_to_rm.each { |p| FileUtils.rm_rf(p) }
+
+      unless r == 0
+        fail 'EnergyPlus returned a non-zero exit code. Check the stdout-energyplus log.'
+      end
+
+      # TODO: check the end or err file
+      if File.exist? 'eplusout.err'
+        eplus_err = File.read('eplusout.err')
+        eplus_err = eplus_err.force_encoding('ISO-8859-1').encode('utf-8', replace: nil)
+        if eplus_err =~ /EnergyPlus Terminated--Fatal Error Detected/
+          fail 'EnergyPlus Terminated with a Fatal Error. Check eplusout.err log.'
+        end
+      end
     rescue => e
       log_message = "#{__FILE__} failed with #{e.message}, #{e.backtrace.join("\n")}"
       @logger.error log_message
+      raise log_message
     ensure
       Dir.chdir(current_dir)
       @logger.info 'EnergyPlus Completed'