+++ spec/files/mongoid/models/measure.rb
@@ -3,7 +3,7 @@ class Measure
   include Mongoid::Timestamps
 
   field :uuid, type: String
-  field :_id, type: String, default: -> { uuid || UUID.generate }
+  field :_id, type: String, default: -> { uuid || SecureRandom.uuid }
   field :version_uuid, type: String # pointless at this time
   field :name, type: String
   field :display_name, type: String
@@ -11,7 +11,7 @@ class Measure
   field :arguments # This is really the variable? right?
   field :measure_type, type: String
   field :values, type: Array, default: []
-  field :index, type: Integer  # how do we set the index, i guess an oncreate call back :~
+  field :index, type: Integer # how do we set the index, i guess an oncreate call back :~
 
   # Relationships
   belongs_to :analysis
@@ -23,27 +23,67 @@ class Measure
   index(name: 1)
   index(analysis_id: 1)
   index(analysis_id: 1, uuid: 1)
-
-  # Validations
-  # validates_format_of :uuid, :with => /[^0-]+/
-  # validates_attachment :seed_zip, content_type: { content_type: "application/zip" }
+  index({analysis_id: 1, name: 1}, unique: true)
 
   # Callbacks
   after_create :verify_uuid
 
-  # parse openstudio's json to get out the variables
+  # Parse Analysis JSON to pull out the measures and variables
+  # Format of JSON is typically
+  # {
+  #     "measure_definition_class_name": "AddOverhangsByProjectionFactor",
+  #     "measure_definition_directory": "./measures/AddOverhangsByProjectionFactor",
+  #     "measure_definition_display_name": "AddOverhangsByProjectionFactor",
+  #     "measure_definition_uuid": "a15669c4-2d37-4177-913c-f62bd8159c7a",
+  #     "measure_definition_version_uuid": "1de14091-ccb5-4dfd-8476-a3e4873af54b",
+  #     "measure_type": "RubyMeasure",
+  #     "arguments": [
+  #       {
+  #         "display_name": "Cardinal Direction",
+  #         "display_name_short": "Cardinal Direction",
+  #         "name": "facade",
+  #         "value": "East",
+  #         "value_type": "choice"
+  #       }
+  #     ]
+  #     "display_name": "Overhangs PF East",
+  #     "name": "overhangs_pf_east",
+  #     "variables": [
+  #       {
+  #         "argument": {
+  #           "display_name": "East Projection Factor",
+  #           "display_name_short": "East PF",
+  #           "name": "projection_factor",
+  #           "value_type": "double"
+  #         },
+  #         "display_name": "East Projection Factor",
+  #         "display_name_short": "East PF",
+  #         "maximum": 1.0,
+  #         "minimum": 0.0,
+  #         "relation_to_output": "",
+  #         "static_value": 0.0,
+  #         "uncertainty_description": {}
+  #       }
+  #     ],
+  #     "workflow_index": 8,
+  #     "workflow_step_type": "Measure",
+  #     "uuid": "657880ff-8a11-4e8a-ac71-935f26f89e48",
+  #     "version_uuid": "8fddf44c-3335-4908-a824-042eec210984"
+  #   }
   def self.create_from_os_json(analysis_id, os_json, pat_json)
-    # The UUID is a misnomer because in measure groups the exact same measure
-    # is copied over multiple times.  The BCL UUID is the actual unique ID IMO.
-    measure = Measure.where(analysis_id: analysis_id, uuid: os_json['uuid']).first
+    # The UUID is a misnomer because in measure groups the exact same measure is copied over multiple times. The
+    # Unique ID is the anlaysis id and the name combined.
+    # **Measure names must be unique and enforced by whatever software is writing the JSON**
+    measure = Measure.where(analysis_id: analysis_id, name: os_json['name']).first
     if measure
-      Rails.logger.info("Measure already exists for analysis #{analysis_id} of #{measure.name} : #{measure.uuid}")
+      # since the index is unique, this should fail before this point
+      fail "Measure already exists for analysis #{analysis_id} of #{measure.name}"
     else
-      measure = Measure.find_or_create_by(analysis_id: analysis_id, uuid: os_json['uuid'])
-      Rails.logger.info("Creating new measure for analysis #{analysis_id} with uuid '#{measure.uuid}'")
+      measure = Measure.find_or_create_by(analysis_id: analysis_id, name: os_json['name'])
+      Rails.logger.info("Creating new measure for analysis #{analysis_id} with name '#{measure.name}'")
     end
 
-    Rails.logger.info("adding/updating measure #{measure.uuid} for analysis #{analysis_id}")
+    Rails.logger.info("Adding/updating measure #{measure.name} for analysis #{analysis_id}")
     i_measure = 0
     os_json.each do |k, v|
       exclude_fields = %w(arguments variables)
@@ -58,49 +98,41 @@ class Measure
         Rails.logger.info('Null measure found')
         measure.name = 'NullMeasure'
       end
+    end
 
-      # Also pull out the value fields for each for each of these and save into a variable instance "somehow???"
-      if k == 'arguments'
-        # Rails.logger.info("checking arguments for values")
-        # just append this to an array for now...
-        # jam this data into the measure for now, but this needs to get pulled out into
-        # Rails.logger.info("#{k.inspect} #{v.inspect}")
-        if v
-          v.each do |arg|
-            # Rails.logger.info(arg.inspect)
+    # Pull out the arugments that are in the measure
+    if os_json['arguments']
+      # Rails.logger.info("#{k.inspect} #{v.inspect}")
+      os_json['arguments'].each do |arg|
+        # Create a variable definition (i.e. a variable) for each argument regardless
+        # whether or not it is used
+        new_var = Variable.create_and_assign_to_measure(analysis_id, measure, arg)
+        # Rails.logger.info("New variable is #{new_var}")
+        measure.variables << new_var unless measure.variables.include?(new_var)
 
-            # Create a variable definition (i.e. a variable) for each argument regardless
-            # whether or not it is used
-            new_var = Variable.create_by_os_argument_json(analysis_id, arg)
-            # Rails.logger.info("New variable is #{new_var}")
-            measure.variables << new_var unless measure.variables.include?(new_var)
+        if pat_json
+          # The measure.values field is just a list of all the set values for the
+          # measure groups which really isn't needed for LHS nor optimization.
+          if arg['value'] && arg['argument_index']
+            # let the system know that the variable was selected for "manipulation"
+            # Rails.logger.info("letting the system know that it can use this variable #{new_var.inspect}")
+            # new_var.perturbable = true
+            # new_var.save!
 
-            if pat_json
-              # The measure.values field is just a list of all the set values for the
-              # measure groups which really isn't needed for LHS nor optimization.
-              if arg['value'] && arg['argument_index']
-                # let the system know that the variable was selected for "manipulation"
-                # Rails.logger.info("letting the system know that it can use this variable #{new_var.inspect}")
-                # new_var.perturbable = true
-                # new_var.save!
-
-                # Rails.logger.info("adding #{arg['value']}")
-                measure.values << [arg['argument_index'], arg['value']]
-              end
-            end
+            # Rails.logger.info("adding #{arg['value']}")
+            measure.values << [arg['argument_index'], arg['value']]
           end
         end
       end
+    end
 
-      if k == 'variables'
-        v.each do |json_var|
-          Rails.logger.info "JSON had a variable named '#{json_var[:name]}'"
-          new_var = Variable.create_by_os_argument_json(analysis_id, json_var)
+    if os_json['variables']
+      os_json['variables'].each do |json_var|
+        Rails.logger.info "JSON had a variable named '#{json_var['display_name']}'"
+        new_var = Variable.create_and_assign_to_measure(analysis_id, measure, json_var)
 
-          if new_var.save!
-            measure.variables << new_var  unless measure.variables.include?(new_var)
-          end
-
+        if new_var.save!
+          measure.variables << new_var unless measure.variables.include?(new_var)
         end
       end
     end