+++ spec/openstudio/workflow/workflow_spec.rb
@@ -49,7 +49,7 @@ describe 'OpenStudio::Workflow' do
 
   end
 
-  it 'should run a local file adapater in legacy mode' do
+  it 'should run a local file adapter in legacy mode' do
     # for local, it uses the rundir as the uuid
     run_dir = './spec/files/local_ex1'
     options = {
@@ -85,26 +85,101 @@ describe 'OpenStudio::Workflow' do
     expect(k.final_state).to eq :finished
   end
 
-  it 'should run a local file with pat format' do
+  it 'should run a local file as energyplus only' do
     # for local, it uses the rundir as the uuid
-    run_dir = './spec/files/pat_project/data_point_469b52c3-4aae-4cdd-b580-5c9494eefa11'
-    options = {
-      is_pat: true,
-      problem_filename: '../formulation.json',
-      datapoint_filename: 'data_point.json',
-      analysis_root_path: 'spec/files/pat_project'
-    }
-    k = OpenStudio::Workflow.load 'Local', run_dir, options
+    run_dir = './spec/files/local_ep'
+    FileUtils.mkdir_p run_dir
+
+    k = OpenStudio::Workflow.run_energyplus 'Local', run_dir
+
+    # copy in an idf and epw file
+    FileUtils.copy('spec/files/example_models/seed/seed.idf', run_dir)
+    FileUtils.copy('spec/files/example_models/weather/in.epw', run_dir)
+
     expect(k).to be_instance_of OpenStudio::Workflow::Run
-    expect(k.options[:problem_filename]).to eq '../formulation.json'
-    expect(k.options[:datapoint_filename]).to eq 'data_point.json'
+    expect(k.options[:problem_filename]).to eq nil
+    expect(k.options[:datapoint_filename]).to eq nil
     expect(k.directory).to eq File.expand_path(run_dir)
     expect(k.run).to eq :finished
     expect(k.final_state).to eq :finished
 
-    expect(File.exist?("#{run_dir}/data_point.zip")).to eq true
-    expect(File.exist?("#{run_dir}/data_point_reports.zip")).to eq true
+    # clean everything up
+  end
 
+  it 'should fail to run an invalid energyplus file' do
+    # for local, it uses the rundir as the uuid
+    run_dir = './spec/files/local_ep_bad'
+    FileUtils.mkdir_p run_dir
+
+    k = OpenStudio::Workflow.run_energyplus 'Local', run_dir
+
+    # copy in an idf and epw file
+    FileUtils.copy('spec/files/example_models/seed/seed_malformed.idf', run_dir)
+    FileUtils.copy('spec/files/example_models/weather/in.epw', run_dir)
+
+    expect(k).to be_instance_of OpenStudio::Workflow::Run
+    expect(k.options[:problem_filename]).to eq nil
+    expect(k.options[:datapoint_filename]).to eq nil
+    expect(k.directory).to eq File.expand_path(run_dir)
+    expect(k.run).to eq :errored
+    expect(k.final_state).to eq :errored
+  end
+
+  it 'should fail to run a file that produces utf-8' do
+    # for local, it uses the rundir as the uuid
+    run_dir = './spec/files/local_ep_iso-8859'
+    FileUtils.mkdir_p run_dir
+
+    k = OpenStudio::Workflow.run_energyplus 'Local', run_dir
+
+    # copy in an idf and epw file
+    FileUtils.copy('spec/files/example_models/seed/seed_8859-1.idf', run_dir)
+    FileUtils.copy('spec/files/example_models/weather/in.epw', run_dir)
+
+    expect(k).to be_instance_of OpenStudio::Workflow::Run
+    expect(k.options[:problem_filename]).to eq nil
+    expect(k.options[:datapoint_filename]).to eq nil
+    expect(k.directory).to eq File.expand_path(run_dir)
+    expect(k.run).to eq :finished
+    expect(k.final_state).to eq :finished
+  end
+
+  it 'should fail to run energyplus with no weather' do
+    # for local, it uses the rundir as the uuid
+    run_dir = './spec/files/local_ep_no_weather'
+    FileUtils.mkdir_p run_dir
+
+    k = OpenStudio::Workflow.run_energyplus 'Local', run_dir
+
+    # copy in an idf and epw file
+    FileUtils.copy('spec/files/example_models/seed/seed.idf', run_dir)
+    # FileUtils.copy('spec/files/example_models/weather/in.epw', run_dir)
+
+    expect(k).to be_instance_of OpenStudio::Workflow::Run
+    expect(k.options[:problem_filename]).to eq nil
+    expect(k.options[:datapoint_filename]).to eq nil
+    expect(k.directory).to eq File.expand_path(run_dir)
+    expect(k.run).to eq :errored
+    expect(k.final_message).to match /.*EPW file not found or not sent to RunEnergyplus.*/
+  end
+
+  it 'should fail to run energyplus with malformed weather' do
+    # for local, it uses the rundir as the uuid
+    run_dir = './spec/files/local_ep_malformed_weather'
+    FileUtils.mkdir_p run_dir
+
+    k = OpenStudio::Workflow.run_energyplus 'Local', run_dir
+
+    # copy in an idf and epw file
+    FileUtils.copy('spec/files/example_models/seed/seed.idf', run_dir)
+    FileUtils.copy('spec/files/example_models/weather/in_malformed.epw', run_dir)
+
+    expect(k).to be_instance_of OpenStudio::Workflow::Run
+    expect(k.options[:problem_filename]).to eq nil
+    expect(k.options[:datapoint_filename]).to eq nil
+    expect(k.directory).to eq File.expand_path(run_dir)
+    expect(k.run).to eq :errored
+    expect(k.final_message).to match /.*failed with EnergyPlus Terminated with a Fatal Error.*/
   end
 
   it 'should not find the input file' do
@@ -232,28 +307,14 @@ describe 'OpenStudio::Workflow' do
     expect(File.exist?("#{run_dir}/reports/eplustbl.html")).to eq true
     expect(File.exist?("#{run_dir}/reports/standard_reports.html")).to eq true
     expect(Dir.exist?("#{run_dir}/run/SetWindowToWallRatioByFacade")).to eq false
+
+    # Test the measure attribute writing with pipes and periods
+    expect(File.exist?("#{run_dir}/run/results.json")).to eq true
+    expect(k.job_results[:run_reporting_measures][:rotate_building_relative_to_current_orientation].key?('An Attribute with Spaces'.to_sym)).to eq true
+    expect(k.job_results[:run_reporting_measures][:rotate_building_relative_to_current_orientation].key?('Invalid_Period_Measure'.to_sym)).to eq true
+    expect(k.job_results[:run_reporting_measures][:rotate_building_relative_to_current_orientation].key?('Invalid_Pipe_Measure'.to_sym)).to eq true
+    expect(k.job_results[:run_reporting_measures][:rotate_building_relative_to_current_orientation].key?('Other_Random_Characters_with_dangling'.to_sym)).to eq true
+    expect(k.job_results[:run_reporting_measures][:rotate_building_relative_to_current_orientation].key?('Asterisks_Are_Bad_Too'.to_sym)).to eq true
+    expect(k.job_results[:run_reporting_measures][:rotate_building_relative_to_current_orientation].key?('Every_Bad_Character_Here_Too'.to_sym)).to eq true
   end
-
-  # it 'should add a new state and transition' do
-  #   transitions = OpenStudio::Workflow::Run.default_transition
-  #   transitions[1][:to] = :xml
-  #   transitions.insert(2, {from: :xml, to: :openstudio})
-  #
-  #   states = OpenStudio::Workflow::Run.default_states
-  #   states.insert(2, {:state => :xml, :options => {:after_enter => :run_xml}})
-  #   options = {
-  #       transitions: transitions,
-  #       states: states,
-  #       analysis_root_path: '../assetscore-openstudio/PNNL_Multi_Block_OS_Console/test_measures',
-  #       xml_library_file: '../assetscore-openstudio/PNNL_Multi_Block_OS_Console/main'
-  #   }
-  #   pp options
-  #   run_dir = './spec/files/mongo_xml1'
-  #   k = OpenStudio::Workflow.load 'Local', run_dir, options
-  #   expect(k).to be_instance_of OpenStudio::Workflow::Run
-  #   expect(k.directory).to eq run_dir
-  #   expect(k.run).to eq :finished
-  #   expect(k.final_state).to eq :finished
-  # end
-
 end