+++ CHANGELOG.md
@@ -1,8 +1,23 @@
 OpenStudio::Workflow Change Log
 ==================================
 
-Unreleased
+Version 0.0.5 (Unreleased)
+-------------
+
+Version 0.0.4
+-------------
+* Include rubyXL gem for reading/writing MS Excel files
+* Remove invalid characters from OpenStudio Measure Attributes. /[|!@#\$%^&\*\(\)\{\}\\\[\]|;:'",<.>\/?\+=]+/
+* Fix objective functions to read from any of the results hash, not just the standard report legacy
+* Add time logger and reporting measure
+
+Version 0.0.3
 --------------
+* Allow measures to set weather file in a measure and have it update what EnergyPlus uses for the weather file.
+* OpenStudio::Workflow.run_energyplus method added to just run energyplus.
+* Remove AASM (act as state machine) and replace with simple tracking of the state. Interface is the same except there is now no need to pass in the States.
+* Catch EnergyPlus errors (non-zero exits), Fatal Errors in eplusout.err, and invalid weather files.
+* Force UTF-8 version upon reading the eplusout.err file
 
 Version 0.0.2
 --------------